Refactor cloud-init config
Use yamlencode instead of concatenating blocks of YAML. Reasoning is
that it's more readable and Terraform can deal with the formatting.

Renamed the following variables and changed their types in line with the
yamlencode approach:
- additional_yaml -> additional_cloud_init
- additional_user_data -> additional_runcmd

Currently the same yaml & user_data is used for all control plane and
node servers in this version.