fix: Drop template dependency
With c921f2c003d32bedb85097d4ce63b7ecb88a7313, all usages of the template
provider's `template_file` were replaced with `templatefile`, which is part of
Terraform > 0.12, hence the provider is not required anymore.

Since the template provider is deprecated, it's not available on non-x86ish
architectures, causing the whole terraform-hcloud-k3s module to be unusable on
them.